# docs/swagger.yaml
definitions:
  domain.Conversation:
    properties:
      assigned_agent_id:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      status:
        type: string
      tenant_id:
        type: integer
    type: object
  domain.Tenant:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
  models.TenantCreateRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models.TenantUpdateRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models.UpdateConversationStatusRequest:
    properties:
      status:
        enum:
        - open
        - closed
        - assigned
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
      summary: User login
      tags:
      - auth
  /api/v1/backoffice/tenants:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of tenants with optional filtering by
        name.
      parameters:
      - description: Search term for tenant name
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Tenant'
      summary: Get a list of tenants
      tags:
      - Tenants
    post:
      consumes:
      - application/json
      description: Create a new tenant with the specified name.
      parameters:
      - description: Tenant creation payload
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.TenantCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Tenant'
      summary: Create a new tenant
      tags:
      - Tenants
  /api/v1/backoffice/tenants/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing tenant by its ID.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Delete a tenant
      tags:
      - Tenants
    patch:
      consumes:
      - application/json
      description: Update the name of an existing tenant by its ID.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant update payload
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.TenantUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Tenant'
      summary: Update an existing tenant
      tags:
      - Tenants
...
  /conversations:
    get:
      consumes:
      - application/json
      description: Get paginated list of conversations with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by status (open, closed, assigned)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Conversation'
      summary: Get list of conversations
      tags:
      - Conversations
  /conversations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a conversation by its ID
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Delete conversation
      tags:
      - Conversations
    get:
      consumes:
      - application/json
      description: Get a single conversation by its ID
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Conversation'
      summary: Get conversation by ID
      tags:
      - Conversations
  /conversations/{id}/escalate:
    post:
      consumes:
      - application/json
      description: Escalate a conversation to a ticket.
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Escalate conversation to ticket
      tags:
      - Conversations
  /conversations/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a conversation (open, closed, assigned)
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: New status for the conversation
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.UpdateConversationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Update conversation status
      tags:
      - Conversations
...
swagger: "2.0"